pyproject.toml
[project]
name = "expman"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"

authors = [
    { name = "LVidal95", email = "laurent.vidal@ikmail.com" }
]
dependencies = [
    "omegaconf>=2.3.0",
    "pandas>=2.3.3",
    "pyjson>=1.4.1",
    "pyyaml>=6.0.3",
    "typer>=0.19.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
expman = "expman.cli:app"
